requirements.txt
schwab-py>=1.0.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
python-dotenv>=0.19.0